+++ library/core/src/test/java/com/google/android/exoplayer2/upstream/cache/CacheDataSourceTest.java
@@ -15,7 +15,6 @@
  */
 package com.google.android.exoplayer2.upstream.cache;
 
-import static android.net.Uri.EMPTY;
 import static com.google.android.exoplayer2.C.LENGTH_UNSET;
 import static com.google.android.exoplayer2.upstream.cache.CacheAsserts.assertCacheEmpty;
 import static com.google.common.truth.Truth.assertThat;
@@ -51,14 +50,16 @@ public final class CacheDataSourceTest {
 
   private static final byte[] TEST_DATA = new byte[] {0, 1, 2, 3, 4, 5, 6, 7, 8, 9};
   private static final int MAX_CACHE_FILE_SIZE = 3;
-  private static final String KEY_1 = "key 1";
-  private static final String KEY_2 = "key 2";
 
+  private Uri testDataUri;
+  private String testDataKey;
   private File tempFolder;
   private SimpleCache cache;
 
   @Before
   public void setUp() throws Exception {
+    testDataUri = Uri.parse("test_data");
+    testDataKey = CacheUtil.generateKey(testDataUri);
     tempFolder = Util.createTempDirectory(RuntimeEnvironment.application, "ExoPlayerTest");
     cache = new SimpleCache(tempFolder, new NoOpCacheEvictor());
   }
@@ -116,7 +117,7 @@ public final class CacheDataSourceTest {
     // If the user try to access off range then it should throw an IOException
     try {
       cacheDataSource = createCacheDataSource(false, false);
-      cacheDataSource.open(new DataSpec(Uri.EMPTY, TEST_DATA.length, 5, KEY_1));
+      cacheDataSource.open(new DataSpec(testDataUri, TEST_DATA.length, 5, testDataKey));
       fail();
     } catch (IOException e) {
       // success
@@ -128,7 +129,7 @@ public final class CacheDataSourceTest {
     // Read partial at EOS but don't cross it so length is unknown
     CacheDataSource cacheDataSource = createCacheDataSource(false, true);
     assertReadData(cacheDataSource, true, TEST_DATA.length - 2, 2);
-    assertThat(cache.getContentLength(KEY_1)).isEqualTo(LENGTH_UNSET);
+    assertThat(cache.getContentLength(testDataKey)).isEqualTo(LENGTH_UNSET);
 
     // Now do an unbounded request for whole data. This will cause a bounded request from upstream.
     // End of data from upstream shouldn't be mixed up with EOS and cause length set wrong.
@@ -136,12 +137,16 @@ public final class CacheDataSourceTest {
     assertReadDataContentLength(cacheDataSource, true, true);
 
     // Now the length set correctly do an unbounded request with offset
-    assertThat(cacheDataSource.open(new DataSpec(EMPTY, TEST_DATA.length - 2,
-        LENGTH_UNSET, KEY_1))).isEqualTo(2);
+    assertThat(
+            cacheDataSource.open(
+                new DataSpec(testDataUri, TEST_DATA.length - 2, LENGTH_UNSET, testDataKey)))
+        .isEqualTo(2);
 
     // An unbounded request with offset for not cached content
-    assertThat(cacheDataSource.open(new DataSpec(EMPTY, TEST_DATA.length - 2,
-        LENGTH_UNSET, KEY_2))).isEqualTo(LENGTH_UNSET);
+    assertThat(
+            cacheDataSource.open(
+                new DataSpec(Uri.parse("notCachedUri"), TEST_DATA.length - 2, LENGTH_UNSET, null)))
+        .isEqualTo(LENGTH_UNSET);
   }
 
   @Test
@@ -157,6 +162,107 @@ public final class CacheDataSourceTest {
     CacheDataSource cacheDataSource = createCacheDataSource(false, false, 0, null);
     assertReadDataContentLength(cacheDataSource, false, false);
     assertCacheEmpty(cache);
+  }
+
+  @Test
+  public void testSwitchToCacheSourceWithReadOnlyCacheDataSource() throws Exception {
+    // Create a fake data source with a 1 MB default data.
+    FakeDataSource upstream = new FakeDataSource();
+    FakeData fakeData = upstream.getDataSet().newDefaultData().appendReadData(1024 * 1024 - 1);
+    // Insert an action just before the end of the data to fail the test if reading from upstream
+    // reaches end of the data.
+    fakeData
+        .appendReadAction(
+            new Runnable() {
+              @Override
+              public void run() {
+                fail("Read from upstream shouldn't reach to the end of the data.");
+              }
+            })
+        .appendReadData(1);
+    // Create cache read-only CacheDataSource.
+    CacheDataSource cacheDataSource =
+        new CacheDataSource(cache, upstream, new FileDataSource(), null, 0, null);
+
+    // Open source and read some data from upstream as the data hasn't cached yet.
+    DataSpec dataSpec = new DataSpec(testDataUri, 0, C.LENGTH_UNSET, testDataKey);
+    cacheDataSource.open(dataSpec);
+    byte[] buffer = new byte[1024];
+    cacheDataSource.read(buffer, 0, buffer.length);
+
+    // Cache the data.
+    // Although we use another FakeDataSource instance, it shouldn't matter.
+    FakeDataSource upstream2 =
+        new FakeDataSource(
+            new FakeDataSource()
+                .getDataSet()
+                .newDefaultData()
+                .appendReadData(1024 * 1024)
+                .endData());
+    CacheUtil.cache(dataSpec, cache, upstream2, null);
+
+    // Read the rest of the data.
+    while (true) {
+      if (cacheDataSource.read(buffer, 0, buffer.length) == C.RESULT_END_OF_INPUT) {
+        break;
+      }
+    }
+    cacheDataSource.close();
+  }
+
+  @Test
+  public void testSwitchToCacheSourceWithNonBlockingCacheDataSource() throws Exception {
+    // Create a fake data source with a 1 MB default data.
+    FakeDataSource upstream = new FakeDataSource();
+    FakeData fakeData = upstream.getDataSet().newDefaultData().appendReadData(1024 * 1024 - 1);
+    // Insert an action just before the end of the data to fail the test if reading from upstream
+    // reaches end of the data.
+    fakeData
+        .appendReadAction(
+            new Runnable() {
+              @Override
+              public void run() {
+                fail("Read from upstream shouldn't reach to the end of the data.");
+              }
+            })
+        .appendReadData(1);
+
+    // Lock the content on the cache.
+    SimpleCacheSpan cacheSpan = cache.startReadWriteNonBlocking(testDataKey, 0);
+    assertThat(cacheSpan).isNotNull();
+    assertThat(cacheSpan.isHoleSpan()).isTrue();
+
+    // Create non blocking CacheDataSource.
+    CacheDataSource cacheDataSource = new CacheDataSource(cache, upstream, 0);
+
+    // Open source and read some data from upstream without writing to cache as the data is locked.
+    DataSpec dataSpec = new DataSpec(testDataUri, 0, C.LENGTH_UNSET, testDataKey);
+    cacheDataSource.open(dataSpec);
+    byte[] buffer = new byte[1024];
+    cacheDataSource.read(buffer, 0, buffer.length);
+
+    // Unlock the span.
+    cache.releaseHoleSpan(cacheSpan);
+    assertCacheEmpty(cache);
+
+    // Cache the data.
+    // Although we use another FakeDataSource instance, it shouldn't matter.
+    FakeDataSource upstream2 =
+        new FakeDataSource(
+            new FakeDataSource()
+                .getDataSet()
+                .newDefaultData()
+                .appendReadData(1024 * 1024)
+                .endData());
+    CacheUtil.cache(dataSpec, cache, upstream2, null);
+
+    // Read the rest of the data.
+    while (true) {
+      if (cacheDataSource.read(buffer, 0, buffer.length) == C.RESULT_END_OF_INPUT) {
+        break;
+      }
+    }
+    cacheDataSource.close();
   }
 
   private void assertCacheAndRead(boolean unboundedRequest, boolean simulateUnknownLength)
@@ -179,8 +285,10 @@ public final class CacheDataSourceTest {
       boolean unboundedRequest, boolean unknownLength) throws IOException {
     int length = unboundedRequest ? C.LENGTH_UNSET : TEST_DATA.length;
     assertReadData(cacheDataSource, unknownLength, 0, length);
-    assertWithMessage("When the range specified, CacheDataSource doesn't reach EOS so shouldn't "
-        + "cache content length").that(cache.getContentLength(KEY_1))
+    assertWithMessage(
+            "When the range specified, CacheDataSource doesn't reach EOS so shouldn't "
+                + "cache content length")
+        .that(cache.getContentLength(testDataKey))
         .isEqualTo(!unboundedRequest ? C.LENGTH_UNSET : TEST_DATA.length);
   }
 
@@ -190,7 +298,7 @@ public final class CacheDataSourceTest {
     if (length != C.LENGTH_UNSET) {
       testDataLength = Math.min(testDataLength, length);
     }
-    assertThat(cacheDataSource.open(new DataSpec(EMPTY, position, length, KEY_1)))
+    assertThat(cacheDataSource.open(new DataSpec(testDataUri, position, length, testDataKey)))
         .isEqualTo(unknownLength ? length : testDataLength);
 
     byte[] buffer = new byte[100];